requirements.txt
graphql-core>=2.1
